v1.3.2 - tests

+++ internal/app/models/order.go
@@ -20,9 +20,9 @@ const (
 
 // LoyalOrder order type from loyal machine
 type LoyalOrder struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float64 `json:"accrual"`
 }
 
 // Order user list
